Alteração do tipo dos parâmetros

+++ src/main.rs
@@ -1,7 +1,7 @@
 use enigo::{Enigo, MouseControllable, MouseButton};
 
 
-fn mouse_click(x: u32, y: u32){
+fn mouse_click(x: u16, y: u16){
     let mut enigo: Enigo = Enigo::new();
 
     // Movendo o mouse para uma posição específica (x, y)
